[Fix] Napalm drivers (speed float)

+++ napalm/iosxr_netconf/iosxr_netconf.py
@@ -338,7 +338,7 @@ class IOSXRNETCONFDriver(NetworkDriver):
             "vendor": "Cisco",
             "os_version": "",
             "hostname": "",
-            "uptime": -1,
+            "uptime": -1.0,
             "serial_number": "",
             "fqdn": "",
             "model": "",
@@ -365,7 +365,7 @@ class IOSXRNETCONFDriver(NetworkDriver):
 
         # Retrieves uptime
         uptime = napalm.base.helpers.convert(
-            int,
+            float,
             self._find_txt(
                 facts_rpc_reply_etree,
                 ".//suo:system-time/\
@@ -373,7 +373,7 @@ class IOSXRNETCONFDriver(NetworkDriver):
                 default="",
                 namespaces=C.NS,
             ),
-            -1,
+            -1.0,
         )
 
         # Retrieves interfaces name
